+++ amazonia/classes/single_instance.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python3
 
-from troposphere import Ref, Tags, Join, Output, GetAtt, ec2, route53
+from troposphere import Ref, Tags, Join, Output, GetAtt, ec2, route53, Base64
 from amazonia.classes.security_enabled_object import SecurityEnabledObject
 
 
 class SingleInstance(SecurityEnabledObject):
-    def __init__(self, title, vpc, template, keypair, si_image_id, si_instance_type, subnet, instance_dependencies, is_nat=False,
-                 hosted_zone_name=None):
+    def __init__(self, title, vpc, template, keypair, si_image_id, si_instance_type, subnet, instance_dependencies,
+                 iam_instance_profile_arn=None, is_nat=False, hosted_zone_name=None):
         """
         AWS CloudFormation - http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-instance.html
         Troposphere - https://github.com/cloudtools/troposphere/blob/master/troposphere/ec2.py
@@ -21,9 +21,46 @@ class SingleInstance(SecurityEnabledObject):
         :param is_nat: a boolean that is used to determine if the instance will be a NAT or not. Default: False
         :param hosted_zone_name: A hosted zone name for setting up a Route 53 record set for Jump hosts
         :param instance_dependencies: a list of dependencies to wait for before creating the single instance.
+        :param iam_instance_profile_arn: the ARN for an IAM profile that enables cloudwatch logging.
         """
 
         super(SingleInstance, self).__init__(vpc=vpc, title=title, template=template)
+
+        region = subnet.AvailabilityZone[:-1]
+        userdata = """#cloud-config
+# Capture all cloud-config output into a more readable logfile
+output: {all: '| tee -a /var/log/cloud-init-output.log'}
+# update and install packages, reboot if necessary
+package_upgrade: true
+package_reboot_if_required: true
+packages:
+ - perl-Switch
+ - perl-DateTime
+ - perl-Sys-Syslog
+ - perl-LWP-Protocol-https
+
+write_files:
+ - path: /etc/awslogs.cfg
+   content: |
+    [general]
+    state_file = /var/awslogs/state/agent-state
+
+    [/var/log/messages]
+    file = /var/log/messages
+    log_group_name = /var/log/messages
+    log_stream_name = {instance_id}
+    datetime_format = %b %d %H:%M:%S
+
+runcmd:
+# cloudwatch monitoring scripts
+ - curl -so /tmp/CloudWatchMonitoringScripts-1.2.1.zip http://aws-cloudwatch.s3.amazonaws.com/downloads/CloudWatchMonitoringScripts-1.2.1.zip
+ - unzip -d /opt /tmp/CloudWatchMonitoringScripts-1.2.1.zip
+ - echo '*/5 * * * * root /opt/aws-scripts-mon/mon-put-instance-data.pl --mem-util --mem-used --mem-avail --disk-space-util --disk-path=/ --from-cron' > /etc/cron.d/cloudwatch
+# cloudwatch logs agent and forwarding config
+ - curl https://s3.amazonaws.com/aws-cloudwatch/downloads/latest/awslogs-agent-setup.py -O
+ - chmod +x ./awslogs-agent-setup.py
+ - ./awslogs-agent-setup.py -n -r """ + region + """ -c /etc/awslogs.cfg
+"""
 
         self.single = self.template.add_resource(
                            ec2.Instance(
@@ -44,8 +81,12 @@ class SingleInstance(SecurityEnabledObject):
                                # http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-instance.html#cfn-ec2-instance-sourcedestcheck
                                SourceDestCheck=False if is_nat else True,
                                Tags=Tags(Name=Join('', [Ref('AWS::StackName'), '-', title])),
-                               DependsOn=instance_dependencies
-                           ))
+                               DependsOn=instance_dependencies,
+                               UserData=Base64(userdata)
+                               ))
+
+        if iam_instance_profile_arn:
+            self.single.IamInstanceProfile = iam_instance_profile_arn.split('/')[1]
 
         if self.single.SourceDestCheck == 'true':
             self.si_output(nat=False, subnet=subnet)