+++ amazonia/classes/stack.py
@@ -10,7 +10,7 @@ from amazonia.classes.database_unit import DatabaseUnit
 class Stack(object):
     def __init__(self, stack_title, code_deploy_service_role, keypair, availability_zones, vpc_cidr, home_cidrs,
                  public_cidr, jump_image_id, jump_instance_type, nat_image_id, nat_instance_type, autoscaling_units,
-                 database_units, stack_hosted_zone_name):
+                 database_units, stack_hosted_zone_name, iam_instance_profile_arn):
         """
         Create a vpc, nat, jumphost, internet gateway, public/private route tables, public/private subnets
          and collection of Amazonia units
@@ -34,6 +34,7 @@ class Stack(object):
         :param database_units: list of dabase_unit dicts (db_instance_type, db_engine, db_port)
         :param stack_hosted_zone_name: A string containing the name of the Route 53 hosted zone to create record
         sets in.
+        :param iam_instance_profile_arn: the ARN for an IAM instance profile that enables cloudtrail access for logging
         """
         super(Stack, self).__init__()
         self.title = stack_title
@@ -47,6 +48,7 @@ class Stack(object):
         self.hosted_zone_name = stack_hosted_zone_name
         self.autoscaling_units = autoscaling_units if autoscaling_units else []
         self.database_units = database_units if database_units else []
+        self.iam_instance_profile_arn = iam_instance_profile_arn
         self.units = {}
         self.private_subnets = []
         self.public_subnets = []
@@ -114,10 +116,12 @@ class Stack(object):
             vpc=self.vpc,
             template=self.template,
             hosted_zone_name=self.hosted_zone_name,
-            instance_dependencies=self.gateway_attachment.title
+            instance_dependencies=self.gateway_attachment.title,
+            iam_instance_profile_arn=self.iam_instance_profile_arn
         )
 
         [self.jump.add_ingress(sender=home_cidr, port='22') for home_cidr in self.home_cidrs]
+        self.jump.add_egress(receiver=self.public_cidr, port='-1')
 
         self.nat = SingleInstance(
             title=self.title + 'Nat',
@@ -128,7 +132,8 @@ class Stack(object):
             vpc=self.vpc,
             template=self.template,
             is_nat=True,
-            instance_dependencies=self.gateway_attachment.title
+            instance_dependencies=self.gateway_attachment.title,
+            iam_instance_profile_arn=self.iam_instance_profile_arn
         )
 
         # Add Routes